CRM_CiviTournament_Form_Person extends CRM_CiviTournament_Form_Contact
